+++ tests/checks_test.rs
@@ -4,21 +4,25 @@ use libcnb_test::{PackResult, TestRunner, assert_contains};
 
 #[test]
 #[ignore = "integration test"]
-fn checks_reject_pythonhome_env_var() {
+fn checks_reject_forbidden_env_vars() {
     let mut config = default_build_config("tests/fixtures/pyproject_toml_only");
     config.env("PYTHONHOME", "/invalid");
+    config.env("VIRTUAL_ENV", "/invalid");
     config.expected_pack_result(PackResult::Failure);
 
     TestRunner::default().build(config, |context| {
         assert_contains!(
             context.pack_stdout,
             indoc! {"
-                [Error: Unsafe environment variable found]
-                The environment variable `PYTHONHOME` is set, however, it can
-                cause problems with the build so we don't allow using it.
+                [Error: Unsafe environment variable(s) found]
+                The following environment variable(s) can cause problems
+                with the build so we don't allow using them:
 
-                You must unset that environment variable. If you didn't set it
-                yourself, check that it wasn't set by an earlier buildpack.
+                PYTHONHOME
+                VIRTUAL_ENV
+
+                You must unset the above env var(s). If you didn't set them
+                yourself, check if they were set by an earlier buildpack.
 
                 ERROR: failed to build: exit status 1
             "}